[draft][SYCLomatic] Move detail namespace code into seprate files
